% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_list_elements.R
\name{check_list_elements}
\alias{check_list_elements}
\title{Check list elements}
\usage{
check_list_elements(list, ...)
}
\arguments{
\item{list}{List object.}

\item{\dots}{Elements that should be checked for presence.}
}
\value{
Nothing
}
\description{
Check list elements. Unexported function, documented for reference.
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, March 2017
}
