% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demPal.R
\name{demPal}
\alias{demPal}
\alias{rivPal}
\alias{rivPal}
\title{DEM and river color palettes}
\usage{
demPal(logbase = 1.02, n = 100, ...)

rivPal(logbase = 1, n = 150, ...)
}
\arguments{
\item{logbase}{Numerical: stretch colors so that differences in small values
become more apparent. 1 to use original \code{\link{terrain.colors}}.
DEFAULT: 1.02 (demPal), 1 (rivPal)}

\item{n}{Number of colors to be used in palette.
DEFAULT: 100 (demPal), 150 (rivPal)}

\item{\dots}{Further arguments passed to \code{berryFunctions::\link{seqPal}}}
}
\value{
color Palette
}
\description{
DEM and river color palettes
}
\examples{
image(volcano, col=rivPal() )
op <- par(mfcol=c(3,3), mar=c(0,0.3,0,0.3), ann=FALSE)
for(l in seq(0.96, 1.04, 0.01))
   {image(volcano, col=demPal(l), axes=FALSE)
   legend("topright", legend=l)}

}
\seealso{
\code{\link{seqPal}}, \code{\link{terrain.colors}}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Mar 2017
}
\keyword{aplot}
