% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_nc_film.R
\docType{data}
\name{ffberry}
\alias{ffberry}
\title{ffmpeg path on berry's computers}
\format{An object of class \code{character} of length 1.}
\usage{
ffberry
}
\description{
ffmpeg path on berry's computers
}
\seealso{
\code{\link{vis_nc_film}}, where it is defined
}
\keyword{datasets}
