% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ncPoint.R
\name{get_ncPoint}
\alias{get_ncPoint}
\title{Get coordinates of point in nc plot}
\usage{
get_ncPoint(nc, n = 1, plot = FALSE, ...)
}
\arguments{
\item{nc}{nc object from \code{\link{read_nc}}}

\item{n}{Maximum number of poins to be identified. DEFAULT: 1}

\item{plot}{Label the identified points? DEFAULT: FALSE}

\item{\dots}{Further arguments passed to \code{\link{identify}}}
}
\value{
data.frame with \code{n} rows and 3 columns (index, row, col)
}
\description{
Get coordinates of a point (or several points) in an nc plot
}
\examples{
# to be added

}
\seealso{
\code{\link{read_nc}}, \code{\link{vis_nc}}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Feb 2017
}
\keyword{iplot}
