% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/png_pdf.R
\name{png_pdf}
\alias{png_pdf}
\title{graph output file management}
\usage{
png_pdf(file, pdf = FALSE, png = FALSE, width = 7, height = 5,
  units = "in", res = 500, ...)
}
\arguments{
\item{file}{filename without pdf/png extension. Files will never be overwritten,
_1 will be appended, see \code{\link[berryFunctions]{newFilename}}.}

\item{pdf}{Write plots created after calling this function into PDF
instead of R graphics device?}

\item{png}{Ditto for PNG. Set either to TRUE, but not both (will stop with an error).}

\item{width, height}{Graph dimensions.}

\item{units, res}{Graph quality arguments passed only to \code{\link{png}}.}

\item{\dots}{Further arguments passed to \code{\link{pdf}} or \code{\link{png}}}
}
\value{
Nothing, but opens a pdf or png device if either pdf or png is TRUE.
}
\description{
graph output file management.
Do not forget to close device again with code like \code{if(pdf|png) dev.off()}
}
\examples{

png_pdf("dummyplot", pdf=TRUE)
plot(cumsum(rnorm(1000)), type="l")
dev.off()

unlink("dummyplot.pdf")

}
\seealso{
\code{\link{pdf}}, \code{\link{png}}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Feb 2017
}
\keyword{file}
