% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_asc.R
\name{read_asc}
\alias{read_asc}
\title{Read asc files}
\usage{
read_asc(file = file.choose(), proj = NA, ...)
}
\arguments{
\item{file}{Name of file. DEFAULT: \code{\link{file.choose}()}}

\item{proj}{Current projection string or NA.
If not NA, coordinates will be converted to lat-long coordinates.
If \code{proj} is a number, e.g. 3035,
it will be replaced with the online epsg proj4 from
\url{http://www.spatialreference.org/ref/epsg/3035/}.
Note that setting \code{proj} may increase computing time a lot.
\code{proj} is passed as the \code{from} argument to
\code{OSMscale::\link[OSMscale]{projectPoints}}.
DEFAULT: NA}

\item{\dots}{Further arguments passed to \code{\link{read.table}}}
}
\value{
Invisible list with matrices for "asc", "x" and "y",
as well as character strings for the filename ("file" and "name"),
and the projection ("proj").
There are also values for the NA character string ("NAS") and "cellsize",
as well as the metainformation lines from the file ("meta").\cr#'
x and y are original coordinates, unless \code{proj != NA},
in which case they are transformed to lat-long degrees.
}
\description{
Read mHM-typical ASC raster files and create a full set of (projected) coordinates
}
\examples{
# to be added

}
\seealso{
\code{\link{vis_asc}} for plotting the result
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Feb-March 2017
}
\keyword{file}
