% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dem.R
\name{read_dem}
\alias{read_dem}
\title{Read dem and facc files}
\usage{
read_dem(inpath = tcltk::tk_choose.dir(), proj = NA, quiet = FALSE, ...)
}
\arguments{
\item{inpath}{Directory containing mHM input files "dem.asc" and "facc.asc".
DEFAULT: \code{tcltk::\link{tk_choose.dir}()}}

\item{proj}{projection passed to \code{\link{read_asc}}. DEFAULT: NA}

\item{quiet}{Logical: should progress messages be suppressed? DEFAULT: FALSE}

\item{\dots}{Further arguments passed to \code{\link{read_asc}}.}
}
\value{
list as in \code{\link{read_asc}}, but "asc" is replaced with
elements "dem", "facc" and "fdir". The element "folder" is added as well.
}
\description{
Read mHM-typical elevation input data: dem (Digital Elevation Model) and
facc (Flow Accumulation, rivers modeled from DEM)
}
\examples{
# to be added

}
\seealso{
\code{\link{vis_dem}}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Feb-March 2017
}
\keyword{hfile}
