% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_nc.R
\name{read_nc}
\alias{read_nc}
\title{Read NETCDF file}
\usage{
read_nc(file = file.choose(), var = "", lat = "lat", lon = "lon", ...)
}
\arguments{
\item{file}{Name of nc file. May include relative or absolute path.
DEFAULT: \code{\link{file.choose}()}}

\item{var}{Name of variable to extract, interactive choice if \code{var} not
available and there are more than 1 variables in the file. DEFAULT: ""}

\item{lat, lon}{Name of latitude and longitude variables. DEFAULT: "lat","lon"}

\item{\dots}{Further arguments passed to \code{ncdf4::\link[ncdf4]{nc_open}},
like verbose=TRUE}
}
\value{
List with time, lat, lon, var, varname, file, cdf (result of
\code{ncdf4::\link[ncdf4]{nc_open}}, from which the previous elements are extracted.)
}
\description{
Read NETCDF file and extract time, location and a variable from it.
}
\section{Warning}{
 This has not yet been tested on many nc files! Please report any errors.
}

\examples{
# to be added

}
\seealso{
\code{\link{vis_nc}}, \code{\link{get_ncPoint}}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Feb 2017
}
\keyword{file}
\keyword{list}
\keyword{spatial}
