% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riverlines.R
\name{riverlines}
\alias{riverlines}
\title{draw river lines}
\usage{
riverlines(dem, sel, col = rivPal(n = 150), lwd = 1:6, add = FALSE, ...)
}
\arguments{
\item{dem}{List returned by \code{\link{read_dem}}}

\item{sel}{Segment indices returned by \code{\link{vis_river}}}

\item{col}{Color scale. DEFAULT: \code{\link{rivPal}(n=150)}}

\item{lwd}{Line width range. DEFAULT: 1:6}

\item{add}{Logical: add to existing plot? DEFAULT: FALSE}

\item{\dots}{Further arguments passed to \code{\link{segments}}.}
}
\value{
catchment area of each cell
}
\description{
Draw river lines, used within \code{\link{vis_river}}.
}
\examples{
# see vis_river

}
\seealso{
\code{\link{vis_river}}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Mar 2017
}
\keyword{aplot}
