% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_dem.R
\name{vis_dem}
\alias{vis_dem}
\title{Visualize dem and facc files}
\usage{
vis_dem(dem, proj = NA, col = demPal(1.02), rivcol = rivPal(n = 150),
  rasterriv = FALSE, prop = 0.98, pch = 15, cex = c(0.14, 0.25),
  add = FALSE, legend = TRUE, title = "Elevation", quiet = FALSE,
  rivargs = NULL, ...)
}
\arguments{
\item{dem}{List returned by \code{\link{read_dem}} or character string
giving a directory (passed to read_dem).}

\item{proj}{Projection passed to \code{\link{read_asc}}. DEFAULT: NA}

\item{col}{DEM color palette. DEFAULT: \code{\link{demPal}(1.02)}}

\item{rivcol}{River color palette. NA will be prepended if rasterriv=TRUE.
DEFAULT: \code{\link{rivPal}(n=150)}}

\item{rasterriv}{Logical: should river be added as raster (with \code{\link{vis_asc}})
instead of lines (with \code{\link{vis_river}})?
Useful to examine single cells exactly. DEFAULT: FALSE}

\item{prop}{Proportion of non-river, see \code{\link{vis_river}}.
If rasterriv=TRUE, adjust river length with the number of
colors in \code{rivcol}. DEFAULT prop: 0.98}

\item{pch, cex}{Point type + size for \code{\link{colPoints}}.
Only used if \code{dem$proj} != NA.
Can each be two values (DEM + FACC separately).
DEFAULT: pch=15, cex=c(0.14, 0.25)}

\item{add}{Logical: add to existing plot? DEFAULT: FALSE}

\item{legend}{Logical: add \code{\link{colPointsLegend}}? DEFAULT: TRUE}

\item{title}{Character string for legend title. DEFAULT: "Elevation"}

\item{quiet}{Logical: should progress messages be suppressed? DEFAULT: FALSE}

\item{rivargs}{List of arguments passed to
\code{\link{vis_asc}} or \code{\link{vis_river}} for FACC.}

\item{\dots}{Further arguments passed to \code{\link{vis_asc}} for DEM, like overwrite=TRUE.}
}
\value{
Invisible list dem or \code{\link{read_dem}} result
}
\description{
Visualize mHM elevation input data: dem (Digital Elevation Model) and
facc (Flow Accumulation, rivers modeled from DEM)
}
\examples{
# to be added

}
\seealso{
\code{\link{read_dem}}, \code{\link{vis_asc}} for plotting any asc file
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Feb-March 2017
}
\keyword{color}
\keyword{hplot}
\keyword{spatial}
