% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_nc.R
\name{vis_nc}
\alias{vis_nc}
\title{Visualize NETCDF file}
\usage{
vis_nc(nc, vals = nc$var[, , 1], zlab = deparse(substitute(vals)),
  add = FALSE, pch = 15, cex = 1.25, col = seqPal(100, colors =
  RColorBrewer::brewer.pal(9, "OrRd")), x = as.vector(nc$lon),
  y = as.vector(nc$lat), z = as.vector(vals), legargs = list(bg = NA),
  ...)
}
\arguments{
\item{nc}{nc object from \code{\link{read_nc}}}

\item{vals}{Subset of \code{nc$var}iable values to be plotted.
DEFAULT: first slice over time.}

\item{zlab}{Legend title. DEFAULT: deparse(substitute(vals))}

\item{add}{Logical: add to existing plot? DEFAULT: FALSE}

\item{pch, cex}{Point character and character expansion (symbol size). DEFAULT: 15, 1.25}

\item{col}{Color palette. DEFAULT: \code{berryFunctions::\link[berryFunctions]{seqPal}}
using colors from \code{RColorBrewer::\link[RColorBrewer]{brewer.pal}(9, "OrRd")}}

\item{x, y, z}{x,y,z Coordinates. DEFAULT: nc elements lon and lat, \code{vals}}

\item{legargs}{List of arguments passed to
\code{berryFunctions::\link[berryFunctions]{colPointsLegend}}}

\item{\dots}{Further arguments passed to
\code{berryFunctions::\link[berryFunctions]{colPoints}}}
}
\value{
\code{\link{colPoints}} list output
}
\description{
Visualize NETCDF object
}
\examples{
# to be added

}
\seealso{
\code{\link{read_nc}}, \code{\link{colPoints}}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Feb 2017
}
\keyword{hplot}
