% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_nc_all.R
\name{vis_nc_all}
\alias{vis_nc_all}
\title{Visualize NETCDF file}
\usage{
vis_nc_all(nc, index = seq_along(nc$time), Range = range(nc$var, na.rm =
  TRUE), ...)
}
\arguments{
\item{nc}{nc object from \code{\link{read_nc}}}

\item{index}{Integer: time slice numbers to be plotted. Will be truncated
to values actually available. If \code{index} is very long,
please make sure to be directing the output into a PDF, movie or similar.
A progress bar will be displayed to estimate remaining time.
DEFAULT: seq_along(nc$time)}

\item{Range}{Range of values, passed to \code{\link{colPoints}}. Use
\code{Range=NULL} to have an individually fitted range for each time slice.
DEFAULT: range(nc$var, na.rm=TRUE)}

\item{\dots}{Further arguments passed to \code{\link{vis_nc}}.
You likely do not want to mess with z!}
}
\value{
list of \code{\link{colPoints}} list outputs
}
\description{
Visualize several (or even all) time slices of a NETCDF object
}
\examples{
# to be added

}
\seealso{
\code{\link{vis_nc}}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Feb 2017
}
\keyword{hplot}
