% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_river.R
\name{vis_river}
\alias{vis_river}
\title{Visialize DEM river as lines (not raster)}
\usage{
vis_river(facc, proj = NA, prop = 0.98,
  zlab = "Catchment area  [1000 km^2]", lwd = 1:3, add = FALSE, ...)
}
\arguments{
\item{facc}{List returned by \code{\link{read_asc}} or character string
giving facc.asc filename (passed to read_asc).}

\item{proj}{projection passed to \code{\link{read_asc}}. DEFAULT: NA}

\item{prop}{Proportion of catchment areas (of each raster cell)
beyond which a channel counts as river. DEFAULT: 0.98}

\item{zlab}{Legend title. DEFAULT: "Catchment area  [1000 km^2]"}

\item{lwd}{Line width range. CURRENTLY MAPPED TO POINT SIZE!! DEFAULT: 1:3}

\item{add}{Logical: add to existing plot? DEFAULT: FALSE}

\item{\dots}{Further arguments passed to \code{\link{colPoints}} except for lines.}
}
\value{
\code{\link{colPoints}} list with segment coordinates
}
\description{
Convert flow accumulation raster to line segments and plot it.
This enables much faster drawing, making it suitable e.g. for drawing small
maps with the river added on top in the corner of plots, e.g. with
\code{berryFunctions::\link{smallPlot}} as in the example.
}
\examples{
# To be added

facc <- read_asc("U:/mHM_shared/basel_6935051/input/morph/facc.asc", proj=3035)
op <- par(no.readonly=TRUE)
vis_asc(facc, png=FALSE)
riv <- vis_river(facc)
par(op)

library(berryFunctions)
logHist(facc$asc*(facc$cellsize/1000)^2, breaks=30) # km^2

# Add rivers as small plot
\dontrun{ ## Not run in CRAN checks to avoid downloading background map
library(OSMscale)
map <- pointsMap(lat=riv$y, lon=riv$x, zoom=7, proj=pll(), plot=F)
par(mar=c(0,0,0,0))
plot(map, removeMargin=FALSE)
vis_river(facc, add=TRUE, lwd=1:3)

par(op)
plot(cumsum(rnorm(300)), type="l") # some gauge properties, e.g.
smallPlot({
        plot(map, removeMargin=FALSE)
        vis_river(facc, add=TRUE, lwd=1:3, legend=FALSE)
        points(8, 47.55, col="red", pch=3, lwd=3, cex=2)
        }, x1=0, x2=0.4, y1=0.77, y2=1, mar=0, border=NA, bg=NA)

} # end dontrun

}
\seealso{
\code{\link{vis_dem}}, \code{\link{read_dem}}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Mar 2017
}
\keyword{aplot}
