% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_asc.R
\name{write_asc}
\alias{write_asc}
\title{Write asc files}
\usage{
write_asc(asc, file = newFilename(asc$file), ...)
}
\arguments{
\item{asc}{List returned by \code{\link{read_asc}}.}

\item{file}{Filename. DEFAULT: \code{berryFunctions::\link{newFilename}(asc$file)}}

\item{\dots}{Further arguments passed to \code{\link{write.table}}.}
}
\value{
Nothing
}
\description{
Write asc file to disc (useful after changing the asc$asc element).
}
\examples{
# to be added

}
\seealso{
\code{\link{read_asc}}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, March 2017
}
\keyword{file}
