% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_asc.R
\name{vis_asc}
\alias{vis_asc}
\title{Visualize asc files}
\usage{
vis_asc(file = file.choose(), pdf = FALSE, png = TRUE,
  col = seqPal(100), bg = "grey", ...)
}
\arguments{
\item{file}{Name of file. DEFAULT: \code{\link{file.choose}()}}

\item{pdf, png}{Save output to disc? See \code{\link{pdf_png}}.
DEFAULT: PDF=FALSE, png=TRUE}

\item{col}{Color palette. DEFAULT:
\code{berryFunctions::\link[berryFunctions]{seqPal}(100)}}

\item{bg}{Graph background color. DEFAULT: "grey"}

\item{\dots}{Further arguments passed to
\code{berryFunctions::\link[berryFunctions]{colPointsLegend}}}
}
\value{
asc matrix
}
\description{
Visualize asc files (commonly used in mHM)
}
\examples{
# to be added

}
\seealso{
\code{\link{vis_dem}}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Feb 2017
}
\keyword{color}
\keyword{hplot}
\keyword{spatial}
