% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_dem.R
\name{vis_dem}
\alias{vis_dem}
\title{Visualize dem and facc files}
\usage{
vis_dem(inpath = choose.dir(), pdf = FALSE, png = TRUE,
  col = terrain.colors(100), bg = "grey", ...)
}
\arguments{
\item{inpath}{Directory containing mHM input files "dem.asc" and "facc.asc".
DEFAULT: \code{\link{choose.dir}()}}

\item{pdf, png}{Save output to disc? See \code{\link{pdf_png}}.
DEFAULT: PDF=FALSE, png=TRUE}

\item{col}{Color palette. DEFAULT: \code{\link{terrain.colors}(100)}}

\item{bg}{Graph background color. DEFAULT: "grey"}

\item{\dots}{Further arguments passed to
\code{berryFunctions::\link[berryFunctions]{colPointsLegend}}}
}
\value{
dem matrix
}
\description{
Visualize mHM elevation input data: dem (Digital Elevation Model) and
facc (Flow Accumulation, rivers modeled from DEM)
}
\examples{
# to be added

}
\seealso{
\code{\link{vis_asc}} for plotting any asc file
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Feb 2017
}
\keyword{color}
\keyword{hplot}
\keyword{spatial}
