% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_dem.R
\name{vis_dem}
\alias{vis_dem}
\title{Visualize dem and facc files}
\usage{
vis_dem(dem, proj = NA, col = terrain.colors(100), nriv = 150,
  rivcol = c(NA, berryFunctions::seqPal(nriv, colors = c("lightblue",
  "darkblue"))), pch = 15, cex = c(0.14, 0.25), add = FALSE,
  legend = TRUE, title = "Elevation", ...)
}
\arguments{
\item{dem}{List returned by \code{\link{read_dem}} or character string
giving a directory (passed to read_dem).}

\item{proj}{projection passed to \code{\link{read_asc}}. DEFAULT: NA}

\item{col}{DEM color palette. DEFAULT: \code{\link{terrain.colors}(100)}}

\item{nriv}{Number of colors for default river palette. A higher number
means more river stretches will be displayed. DEFAULT: 150}

\item{rivcol}{River color palette. DEFAULT: c(NA, seqPal(nriv, ---))}

\item{pch, cex}{Point type + size for colPoints if \code{dem$proj} != NA.
Can each be two values (DEM + FACC separately).
DEFAULT: pch=15, cex=c(0.14, 0.25)}

\item{add}{Logical: add to existing plot? DEFAULT: FALSE}

\item{legend}{Logical: add \code{\link{colPointsLegend}}? DEFAULT: TRUE}

\item{title}{CHaracter string for legend title. DEFAULT: "Elevation"}

\item{\dots}{Further arguments passed to \code{\link{vis_asc}}.}
}
\value{
Invisible list dem or \code{\link{read_dem}} result
}
\description{
Visualize mHM elevation input data: dem (Digital Elevation Model) and
facc (Flow Accumulation, rivers modeled from DEM)
}
\examples{
# to be added

}
\seealso{
\code{\link{read_dem}}, \code{\link{vis_asc}} for plotting any asc file
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Feb-March 2017
}
\keyword{color}
\keyword{hplot}
\keyword{spatial}
