% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_nc.R
\name{vis_nc}
\alias{vis_nc}
\title{Visualize NETCDF file}
\usage{
vis_nc(nc, index = 1, x = as.vector(nc$lon), y = as.vector(nc$lat),
  z = as.vector(nc$var[, , index]), zlab = paste(nc$varname,
  nc$time[index]), add = FALSE, pch = 15, cex = 1.25, col = seqPal(100,
  colors = RColorBrewer::brewer.pal(9, "OrRd")), bg.leg = NA, cex.leg = 1,
  legargs = NULL, ...)
}
\arguments{
\item{nc}{nc object from \code{\link{read_nc}}}

\item{index}{Integer: a single time slice number to be plotted.
To plot several slices, use \code{\link{vis_nc_all}}. DEFAULT: 1}

\item{x, y, z}{x,y,z Coordinates. DEFAULT: nc elements lon, lat, var (@ time index)}

\item{zlab}{Legend title. DEFAULT: varname time}

\item{add}{Logical: add to existing plot? DEFAULT: FALSE}

\item{pch, cex}{Point character and character expansion (symbol size). DEFAULT: 15, 1.25}

\item{col}{Color palette. DEFAULT: \code{berryFunctions::\link[berryFunctions]{seqPal}}
using colors from \code{RColorBrewer::\link[RColorBrewer]{brewer.pal}(9, "OrRd")}}

\item{bg.leg, cex.leg}{Background and character size for legend. DEFAULT: NA, 1}

\item{legargs}{List of further arguments passed to legend (
\code{berryFunctions::\link[berryFunctions]{colPointsLegend}})}

\item{\dots}{Further arguments passed to
\code{berryFunctions::\link[berryFunctions]{colPoints}}}
}
\value{
\code{\link{colPoints}} list output
}
\description{
Visualize a single time slice of a NETCDF object
}
\examples{
# to be added

}
\seealso{
\code{\link{read_nc}},
         \code{\link{vis_nc_all}} and \code{\link{vis_nc_film}} for several time slices,
         \code{\link{get_ncPoint}}, \code{\link{colPoints}}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Feb 2017
}
\keyword{hplot}
