% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_nc_all.R
\name{vis_nc_all}
\alias{vis_nc_all}
\title{Visualize NETCDF file}
\usage{
vis_nc_all(nc, index = 1, Range = range(nc$var, na.rm = TRUE),
  expr = NULL, vlcnote = FALSE, ...)
}
\arguments{
\item{nc}{nc object from \code{\link{read_nc}}}

\item{index}{Integer: time slice numbers to be plotted. Will be truncated
to values actually available. If \code{index} is very long,
please make sure to be directing the output into a PDF, movie or similar.
For automated movie generation, see \code{\link{vis_nc_film}}.
A progress bar will be displayed to estimate remaining time.
DEFAULT: 1}

\item{Range}{Range of values, passed to \code{\link{colPoints}}. Use
\code{Range=NULL} to have an individually fitted range for each time slice.
DEFAULT: range(nc$var, na.rm=TRUE)}

\item{expr}{Expression (potentially several lines wrapped in curly braces)
to be executed after each call of \code{\link{vis_nc}},
like statements using axis, title, mtext etc. Can include
references to current \code{index} value at each time step.
DEFAULT: NULL}

\item{vlcnote}{Add VLC instruction note in each frame? DEFAULT: FALSE}

\item{\dots}{Further arguments passed to \code{\link{vis_nc}}.
You likely do not want to mess with z!}
}
\value{
list of \code{\link{colPoints}} list outputs
}
\description{
Visualize several (or even all) time slices of a NETCDF object
}
\examples{
# to be added

}
\seealso{
\code{\link{vis_nc}}, \code{\link{vis_nc_film}} to create an animated movie
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Feb 2017
}
\keyword{hplot}
